requirements.txt
botbuilder
botbuilder-core
azure-mgmt-resource
azure-mgmt-web
azure-mgmt-consumption
azure-identity
